pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-ksa-ai"
version = "0.1.0"
description = "Local Arabic-first RAG assistant for Faculty AI regulations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pymupdf>=1.23.0",
  "scikit-learn>=1.3.0",
  "faiss-cpu>=1.7.4",
  "numpy>=1.24.0",
  "flask>=3.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
